requirements/ci.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-index --output-file=requirements/ci.txt requirements/base.txt requirements/test-tools.in
#
aiodns==2.0.0             # via -r requirements/base.txt
aiohttp==3.6.1            # via -r requirements/base.txt
amqp==2.5.1               # via -r requirements/base.txt, kombu
async-timeout==3.0.1      # via -r requirements/base.txt, aiohttp
attrs==19.1.0             # via -r requirements/base.txt, aiohttp
beautifulsoup4==4.8.2     # via webtest
billiard==3.6.1.0         # via -r requirements/base.txt, celery
celery==4.3.0             # via -r requirements/base.txt, django-camunda
certifi==2019.9.11        # via -r requirements/base.txt, requests
cffi==1.12.3              # via -r requirements/base.txt, cryptography, pycares
chardet==3.0.4            # via -r requirements/base.txt, aiohttp, requests
coverage==5.0.3           # via -r requirements/test-tools.in
cryptography==2.8         # via -r requirements/base.txt, django-auth-adfs
cssselect==1.1.0          # via pyquery
defusedxml==0.6.0         # via -r requirements/base.txt, django-camunda, odfpy
diff-match-patch==20181111  # via -r requirements/base.txt, django-import-export
django-appconf==1.0.3     # via -r requirements/base.txt, django-axes
django-auth-adfs-db==0.2.0  # via -r requirements/base.txt
django-auth-adfs==1.3.1   # via -r requirements/base.txt, django-auth-adfs-db
django-axes==5.0.13       # via -r requirements/base.txt
django-camunda==0.9.2     # via -r requirements/base.txt
django-choices==1.7.0     # via -r requirements/base.txt, zgw-consumers
django-import-export==1.2.0  # via -r requirements/base.txt
django-ipware==2.1.0      # via -r requirements/base.txt, django-axes
django-redis==4.10.0      # via -r requirements/base.txt
django-relativedelta==1.1.1  # via -r requirements/base.txt, zgw-consumers
django-sniplates==0.6.0   # via -r requirements/base.txt
django-solo==1.1.3        # via -r requirements/base.txt, django-auth-adfs-db, django-camunda, zgw-consumers
django-webtest==1.9.7     # via -r requirements/test-tools.in
django==2.2.12            # via -r requirements/base.txt, django-appconf, django-auth-adfs, django-auth-adfs-db, django-axes, django-camunda, django-choices, django-import-export, django-redis, django-relativedelta, django-sniplates, djangorestframework, zgw-consumers
djangorestframework==3.11.0  # via -r requirements/base.txt
et-xmlfile==1.0.1         # via -r requirements/base.txt, openpyxl
factory-boy==2.12.0       # via -r requirements/test-tools.in
faker==4.0.2              # via factory-boy
freezegun==0.3.15         # via -r requirements/test-tools.in
furl==2.1.0               # via -r requirements/base.txt
gemma-zds-client==0.13.0  # via -r requirements/base.txt, zgw-consumers
idna==2.8                 # via -r requirements/base.txt, requests, yarl
importlib-metadata==0.23  # via -r requirements/base.txt, kombu
inflection==0.3.1         # via -r requirements/base.txt, django-camunda
isodate==0.6.0            # via -r requirements/base.txt
jdcal==1.4.1              # via -r requirements/base.txt, openpyxl
kombu==4.6.4              # via -r requirements/base.txt, celery
lxml==4.5.0               # via pyquery
markuppy==1.14            # via -r requirements/base.txt, tablib
more-itertools==7.2.0     # via -r requirements/base.txt, zipp
multidict==4.5.2          # via -r requirements/base.txt, aiohttp, yarl
odfpy==1.4.0              # via -r requirements/base.txt, tablib
openpyxl==3.0.0           # via -r requirements/base.txt, tablib
orderedmultidict==1.0.1   # via -r requirements/base.txt, furl
psycopg2==2.8.3           # via -r requirements/base.txt
pycares==3.0.0            # via -r requirements/base.txt, aiodns
pycparser==2.19           # via -r requirements/base.txt, cffi
pyjwt==1.7.1              # via -r requirements/base.txt, django-auth-adfs, gemma-zds-client
pyquery==1.4.1            # via -r requirements/test-tools.in
python-dateutil==2.8.0    # via -r requirements/base.txt, django-camunda, django-relativedelta, faker, freezegun
python-decouple==3.3      # via -r requirements/base.txt
python-dotenv==0.10.3     # via -r requirements/base.txt
pytz==2019.2              # via -r requirements/base.txt, celery, django
pyyaml==5.1.2             # via -r requirements/base.txt, gemma-zds-client, tablib
raven==6.10.0             # via -r requirements/base.txt
redis==3.3.8              # via -r requirements/base.txt, django-redis
requests-mock==1.7.0      # via -r requirements/base.txt, -r requirements/test-tools.in
requests==2.22.0          # via -r requirements/base.txt, django-auth-adfs, django-camunda, gemma-zds-client, requests-mock
rules==2.2                # via -r requirements/base.txt
six==1.12.0               # via -r requirements/base.txt, cryptography, django-appconf, freezegun, furl, isodate, orderedmultidict, python-dateutil, requests-mock, webtest
soupsieve==2.0            # via beautifulsoup4
sqlparse==0.3.0           # via -r requirements/base.txt, django
tablib==0.14.0            # via -r requirements/base.txt, django-import-export
tblib==1.6.0              # via -r requirements/test-tools.in
text-unidecode==1.3       # via faker
urllib3==1.25.5           # via -r requirements/base.txt, requests
uwsgi==2.0.18             # via -r requirements/base.txt
vine==1.3.0               # via -r requirements/base.txt, amqp, celery
waitress==1.4.3           # via webtest
webob==1.8.6              # via webtest
webtest==2.0.34           # via django-webtest
xlrd==1.2.0               # via -r requirements/base.txt, tablib
xlwt==1.3.0               # via -r requirements/base.txt, tablib
yarl==1.3.0               # via -r requirements/base.txt, aiohttp
zgw-consumers==0.9.0      # via -r requirements/base.txt
zipp==0.6.0               # via -r requirements/base.txt, importlib-metadata
